Allow gradients and drop shadows to be removed.
I wanted to be able to add a drop shadow on mouseover and remove it again on mouseout.  I noticed this wasn't possible with the library as written, so I added code to save the value of the filter attribute before setting it for the drop shadow and reapply that value when the style is changed.  I applied the same treatment for gradients.