Fix SVG marker initialization.
This fixex SVG marker initialization when a string rather than a URL is provided.  It also adds a fallback SVG document when nothing is specified. The fallback is useful if you want to display SVG-managed text at a location, for instance.

Appending the group to renderer's container rather than parent node fixes a bug when the path is not actually in the DOM when the path is updating.  This happens when calling setStyle on an existing SVG marker or when SVG text is provided as an option rather than a URL.  In both cases, addSVG is called while _path is not in the DOM.